need to implement server_config onto generate_error_response() generate_response looks to be ok now
