+++ includes/http/HTTPHandler.hpp
@@ -3,11 +3,14 @@
 
 #include "socket/TCPSocket.hpp"
 #include "http/HTTPRequest.hpp"
+#include "http/HTTPResponse.hpp"
+#include "server/ServerConfig.hpp"
 #include "utils/smt.hpp"
 
 namespace webserv {
 
 void http_handle(TCPSocket* sock, int client_fd);
+HTTPResponse	generate_response(std::pair<int, std::string> resp);
 
 } // namespace webserv
