+++ includes/http/HTTPResponse.hpp
@@ -7,6 +7,8 @@
 #include <string>
 #include <fstream>
 
+#include "server/ServerConfig.hpp"
+
 #define SP   std::string(" ")
 #define CRLF std::string("\r\n")
 
@@ -21,7 +23,7 @@ struct HTTPResponse {
 
         /* Constructor and Destructor */
         HTTPResponse(void); // init == false
-		HTTPResponse(int status_code);
+		HTTPResponse(int status_code, ServerConfig config);
         HTTPResponse(int code, HTTPHeader header, std::string body); // init == true
 		~HTTPResponse(void);
         // HTTPResponse(HTTPResponse const& src);
@@ -30,6 +32,8 @@ struct HTTPResponse {
         /* Other Methods */
         std::string to_str(void) const;
 		std::map<int, std::string>	create_code_map( void ) const;
+		std::map<std::string, std::string> create_mime_map(void) const;
+
 
         /* Member Attributes */
         bool        init;
