--- srcs/HTTPResponse.cpp
@@ -1,745 +1,0 @@
-#include "HTTPResponse.hpp"
-
-namespace webserv {
-
-HTTPResponse::HTTPResponse(void): init(false) {} 
-
-HTTPResponse::HTTPResponse(int status_code): init(true), status_map(this->create_code_map()), version("HTTP/1.1") {
-	//getting status
-	for (StatusMap::const_iterator it = this->status_map.begin(); it != this->status_map.end(); it++) {
-		if (status_code == it->first) {
-			this->status.first = status_code;
-			this->status.second = it->second;
-			break;
-		}
-	}
-	if (this->status.first == 0) {
-		this->status.first = 500;
-		this->status.second = "Internal Server Error";
-	}
-	//getting body
-	std::stringstream ss;
-    ss << this->status.first;
-    std::string error_code = ss.str();
-	std::string	error_file = "../webserv/error_pages/" + error_code + ".html";
-	std::ifstream	body_file(error_file);
-	if (body_file.fail()) {
-		body_file.close();
-		body_file.open("../webserv/error_pages/500.html");
-	}
-	this->body = std::string((std::istreambuf_iterator<char>(body_file)), std::istreambuf_iterator<char>());
-	if (!this->header.empty())
-		this->header.clear();
-	//getting headers
-	this->header.insert(std::pair<std::string, std::string>("Content-type", "text/html"));
-	ss.str("");
-	ss << body.size();
-	this->header.insert(std::pair<std::string, std::string>("Content-length", ss.str()));
-	
-}
-
-HTTPResponse::HTTPResponse(int code, HTTPHeader header, std::string body): status_map(this->create_code_map()), header(header), body(body), init(true),
-	version("HTTP/1.1") {
-	StatusMap::const_iterator it = this->status_map.find(code);
-	if (it != this->status_map.end()) {
-		this->status.first = code;
-		this->status.second = it->second;
-	}
-	else {
-		this->status.first = 500;
-		this->status.second = "Internal Server Error";
-		//getting body
-		std::ifstream file("../webserv/error_pages/500.html");
-		this->body = std::string((std::istreambuf_iterator<char>(file)), std::istreambuf_iterator<char>());
-		//getting headers
-		std::stringstream	ss;
-		this->header.insert(std::pair<std::string, std::string>("Content-type", "text/html"));
-		ss << this->body.size();
-		this->header.insert(std::pair<std::string, std::string>("Content-length", ss.str()));
-	}
-}
-
-HTTPResponse::~HTTPResponse(void) {
-}
-
-std::string HTTPResponse::to_str(void) const {
-
-    // status line
-    std::stringstream ss;
-    ss << this->status.first;
-    std::string code = ss.str();
-    std::string start_line =
-        std::string(this->version + SP + code + SP + this->status.second);
-
-    // header
-    std::string h = "";
-    for (HTTPHeader::const_iterator it = this->header.begin();
-         it != this->header.end(); it++)
-        h += std::string((*it).first + ": " + (*it).second + CRLF);
-
-    return (std::string(start_line + CRLF + h +
-                        (this->body.empty() ? (CRLF) : (CRLF + body))));
-
-    // return (std::string(start_line /* + CRLF + header.to_str() */ +
-    //                     (this->body.empty() ? (CRLF + CRLF) : (CRLF + CRLF +
-    //                     this->body))));
-}
-
-std::map<int, std::string> HTTPResponse::create_code_map(void) const {
-	std::map<int, std::string>	code_map;
-
-	code_map[100] = "Continue";
-    code_map[101] = "Switching Protocols";
-    code_map[102] = "Processing";
-    code_map[103] = "Early Hints";
-    code_map[200] = "OK";
-    code_map[201] = "Created";
-    code_map[202] = "Accepted";
-    code_map[203] = "Non Authoritative Info";
-    code_map[204] = "No Content";
-    code_map[205] = "Reset Content";
-    code_map[206] = "Partial Content";
-    code_map[207] = "Multi Status";
-    code_map[208] = "Already Reported";
-    code_map[226] = "IM Used";
-    code_map[300] = "Multiple Choices";
-    code_map[301] = "Moved Permanently";
-    code_map[302] = "Found";
-    code_map[303] = "See Other";
-    code_map[304] = "Not Modified";
-    code_map[305] = "Use Proxy";
-    code_map[307] = "Temporary Redirect";
-    code_map[308] = "Permanent Redirect";
-    code_map[400] = "Bad Request";
-    code_map[401] = "Unauthorized";
-    code_map[402] = "Payment Required";
-    code_map[403] = "Forbidden";
-    code_map[404] = "Not Found";
-    code_map[405] = "Method Not Allowed";
-    code_map[406] = "Not Acceptable";
-    code_map[407] = "Proxy Auth Required";
-    code_map[408] = "Request Timeout";
-    code_map[409] = "Conflict";
-    code_map[410] = "Gone";
-    code_map[411] = "Length Required";
-    code_map[412] = "Precondition Failed";
-    code_map[413] = "Request Entity Too Large";
-    code_map[414] = "Request URI Too Long";
-    code_map[415] = "Unsupported Media Type";
-    code_map[416] = "Requested Range Not Satisfiable";
-    code_map[417] = "Expectation Failed";
-    code_map[418] = "Teapot";
-    code_map[421] = "Misdirected Request";
-    code_map[422] = "Unprocessable Entity";
-    code_map[423] = "Locked";
-    code_map[424] = "Failed Dependency";
-    code_map[425] = "Too Early";
-    code_map[426] = "Upgrade Required";
-    code_map[428] = "Precondition Required";
-    code_map[429] = "Too Many Requests";
-    code_map[431] = "Request Header Fields Too Large";
-    code_map[451] = "Unavailable For Legal Reasons";
-    code_map[500] = "Internal Server Error";
-    code_map[501] = "Not Implemented";
-    code_map[502] = "Bad Gateway";
-    code_map[503] = "Service Unavailable";
-    code_map[504] = "GatewayTimeout";
-    code_map[505] = "HTTP Version Not Supported";
-    code_map[506] = "Variant Also Negotiates";
-    code_map[507] = "Insufficient Storage";
-    code_map[508] = "Loop Detected";
-    code_map[510] = "Not Extended";
-    code_map[511] = "Network Authentication Required";
-	return code_map;
-}
-
-// std::map<std::string, std::string> HTTPResponse::create_mime_map(void) {
-// 	std::map<std::string, std::string>	mime_map;
-// 	mime_map["3dm"] =  	"x-world/x-3dmfHTTPResponse";
-// 	mime_map["3dmf"] =  	"x-world/x-3dmf";
-// 	mime_map["a"] =  	"application/octet-stream";
-// 	mime_map["aab"] =  	"application/x-authorware-bin";
-// 	mime_map["aam"] =  	"application/x-authorware-map";
-// 	mime_map["aas"] =  	"application/x-authorware-seg";
-// 	mime_map["abc"] =  	"text/vnd.abc";
-// 	mime_map["acgi"] =  	"text/html";
-// 	mime_map["afl"] =  	"video/animaflex";
-// 	mime_map["ai"] =  	"application/postscript";
-// 	mime_map["aif"] =  	"audio/aiff";
-// 	mime_map["aifc"] =  	"audio/aiff";
-// 	mime_map["aiff"] =  	"audio/aiff";
-// 	mime_map["aiff"] =  	"audio/x-aiff";
-// 	mime_map["aim"] =  	"application/x-aim";
-// 	mime_map["aip"] =  	"text/x-audiosoft-intra";
-// 	mime_map["ani"] =  	"application/x-navi-animation";
-// 	mime_map["aos"] =  	"application/x-nokia-9000-communicator-add-on-software";
-// 	mime_map["aps"] =  	"application/mime";
-// 	mime_map["arc"] =  	"application/octet-stream";
-// 	mime_map["art"] =  	"image/x-jg";
-// 	mime_map["asf"] =  	"video/x-ms-asf";
-// 	mime_map["asm"] =  	"text/x-asm";
-// 	mime_map["asp"] =  	"text/asp";
-// 	mime_map["asx"] =  	"application/x-mplayer2";
-// 	mime_map["asx"] =  	"video/x-ms-asf";
-// 	mime_map["au"] =  	"audio/basic";
-// 	mime_map["avi"] =  	"video/x-msvideo";
-// 	mime_map["avs"] =  	"video/avs-video";
-// 	mime_map["bcpio"] =  	"application/x-bcpio";
-// 	mime_map["bin"] =  	"application/octet-stream";
-// 	mime_map["bm"] =  	"image/bmp";
-// 	mime_map["bmp"] =  	"image/bmp";
-// 	mime_map["boo"] =  	"application/book";
-// 	mime_map["book"] =  	"application/book";
-// 	mime_map["boz"] =  	"application/x-bzip2";
-// 	mime_map["bsh"] =  	"application/x-bsh";
-// 	mime_map["bz"] =  	"application/x-bzip";
-// 	mime_map["bz2"] =  	"application/x-bzip2";
-// 	mime_map["c"] =  	"text/x-c";
-// 	mime_map["cat"] =  	"application/vnd.ms-pki.seccat";
-// 	mime_map["cc"] =  	"text/x-c";
-// 	mime_map["ccad"] =  	"application/clariscad";
-// 	mime_map["cco"] =  	"application/x-cocoa";
-// 	mime_map["cdf"] =  	"application/x-netcdf";
-// 	mime_map["cer"] =  	"application/pkix-cert";
-// 	mime_map["cha"] =  	"application/x-chat";
-// 	mime_map["chat"] =  	"application/x-chat";
-// 	mime_map["class"] =  	"application/java";
-// 	mime_map["com"] =  	"application/octet-stream";
-// 	mime_map["conf"] =  	"text/plain";
-// 	mime_map["cpio"] =  	"application/x-cpio";
-// 	mime_map["cpp"] =  	"text/x-c";
-// 	mime_map["cpt"] =  	"application/mac-compactpro";
-// 	mime_map["crl"] =  	"application/pkix-crl";
-// 	mime_map["crt"] =  	"application/x-x509-ca-cert";
-// 	mime_map["csh"] =  	"application/x-csh";
-// 	mime_map["css"] =  	"text/css";
-// 	mime_map["cxx"] =  	"text/plain";
-// 	mime_map["dcr"] =  	"application/x-director";
-// 	mime_map["deepv"] =  	"application/x-deepv";
-// 	mime_map["def"] =  	"text/plain";
-// 	mime_map["der"] =  	"application/x-x509-ca-cert";
-// 	mime_map["dif"] =  	"video/x-dv";
-// 	mime_map["dir"] =  	"application/x-director";
-// 	mime_map["doc"] =  	"application/msword";
-// 	mime_map["dot"] =  	"application/msword";
-// 	mime_map["dp"] =  	"application/commonground";
-// 	mime_map["drw"] =  	"application/drafting";
-// 	mime_map["dump"] =  	"application/octet-stream";
-// 	mime_map["dv"] =  	"video/x-dv";
-// 	mime_map["dvi"] =  	"application/x-dvi";
-// 	mime_map["dwf"] =  	"model/vnd.dwf";
-// 	mime_map["dwg"] =  	"image/vnd.dwg";
-// 	mime_map["dxf"] =  	"application/dxf";
-// 	mime_map["dxr"] =  	"application/x-director";
-// 	mime_map["el"] =  	"text/x-script.elisp";
-// 	mime_map["env"] =  	"application/x-envoy";
-// 	mime_map["eps"] =  	"application/postscript";
-// 	mime_map["es"] =  	"application/x-esrehber";
-// 	mime_map["etx"] =  	"text/x-setext";
-// 	mime_map["exe"] =  	"application/octet-stream";
-// 	mime_map["f"] =  	"text/x-fortran";
-// 	mime_map["f77"] =  	"text/x-fortran";
-// 	mime_map["f90"] =  	"text/x-fortran";
-// 	mime_map["fdf"] =  	"application/vnd.fdf";
-// 	mime_map["fli"] =  	"video/x-fli";
-// 	mime_map["flo"] =  	"image/florian";
-// 	mime_map["flx"] =  	"text/vnd.fmi.flexstor";
-// 	mime_map["fmf"] =  	"video/x-atomic3d-feature";
-// 	mime_map["for"] =  	"text/x-fortran";
-// 	mime_map["fpx"] =  	"image/vnd.fpx";
-// 	mime_map["frl"] =  	"application/freeloader";
-// 	mime_map["funk"] =  	"audio/make";
-// 	mime_map["g"] =  	"text/plain";
-// 	mime_map["g3"] =  	"image/g3fax";
-// 	mime_map["gif"] =  	"image/gif";
-// 	mime_map["gsd"] =  	"audio/x-gsm";
-// 	mime_map["gsm"] =  	"audio/x-gsm";
-// 	mime_map["gsp"] =  	"application/x-gsp";
-// 	mime_map["gss"] =  	"application/x-gss";
-// 	mime_map["gtar"] =  	"application/x-gtar";
-// 	mime_map["gz"] =  	"application/x-gzip";
-// 	mime_map["h"] =  	"text/x-c";
-// 	mime_map["hdf"] =  	"application/x-hdf";
-// 	mime_map["help"] =  	"application/x-helpfile";
-// 	mime_map["hgl"] =  	"application/vnd.hp-hpgl";
-// 	mime_map["hh"] =  	"text/x-c";
-// 	mime_map["hlb"] =  	"text/x-script";
-// 	mime_map["hlp"] =  	"application/winhelp";
-// 	mime_map["hpg"] =  	"application/vnd.hp-hpgl";
-// 	mime_map["hpgl"] =  	"application/vnd.hp-hpgl";
-// 	mime_map["hqx"] =  	"application/mac-binhex40";
-// 	mime_map["hta"] =  	"application/hta";
-// 	mime_map["htc"] =  	"text/x-component";
-// 	mime_map["htm"] =  	"text/html";
-// 	mime_map["html"] =  	"text/html";
-// 	mime_map["htmls"] =  	"text/html";
-// 	mime_map["htt"] =  	"text/webviewhtml";
-// 	mime_map["htx"] =  	"text/html";
-// 	mime_map["ice"] =  	"x-conference/x-cooltalk";
-// 	mime_map["ico"] =  	"image/x-icon";
-// 	mime_map["idc"] =  	"text/plain";
-// 	mime_map["ief"] =  	"image/ief";
-// 	mime_map["iefs"] =  	"image/ief";
-// 	mime_map["iges"] =  	"application/iges";
-// 	mime_map["iges"] =  	"model/iges";
-// 	mime_map["igs"] =  	"application/iges";
-// 	mime_map["igs"] =  	"model/iges";
-// 	mime_map["ima"] =  	"application/x-ima";
-// 	mime_map["imap"] =  	"application/x-httpd-imap";
-// 	mime_map["inf"] =  	"application/inf";
-// 	mime_map["ins"] =  	"application/x-internett-signup";
-// 	mime_map["ip"] =  	"application/x-ip2";
-// 	mime_map["isu"] =  	"video/x-isvideo";
-// 	mime_map["it"] =  	"audio/it";
-// 	mime_map["iv"] =  	"application/x-inventor";
-// 	mime_map["ivr"] =  	"i-world/i-vrml";
-// 	mime_map["ivy"] =  	"application/x-livescreen";
-// 	mime_map["jam"] =  	"audio/x-jam";
-// 	mime_map["jav"] =  	"text/plain";
-// 	mime_map["jav"] =  	"text/x-java-source";
-// 	mime_map["java"] =  	"text/plain";
-// 	mime_map["java"] =  	"text/x-java-source";
-// 	mime_map["jcm"] =  	"application/x-java-commerce";
-// 	mime_map["jfif"] =  	"image/jpeg";
-// 	mime_map["jfif"] =  	"image/pjpeg";
-// 	mime_map["jfif-tbnl"] = "image/jpeg";
-// 	mime_map["jpe"] =  	"image/jpeg";
-// 	mime_map["jpe"] =  	"image/pjpeg";
-// 	mime_map["jpeg"] =  	"image/jpeg";
-// 	mime_map["jpeg"] =  	"image/pjpeg";
-// 	mime_map["jpg"] =  	"image/jpeg";
-// 	mime_map["jpg"] =  	"image/pjpeg";
-// 	mime_map["jps"] =  	"image/x-jps";
-// 	mime_map["js"] =  	"application/x-javascript";
-// 	mime_map["js"] =  	"application/javascript";
-// 	mime_map["js"] =  	"application/ecmascript";
-// 	mime_map["js"] =  	"text/javascript";
-// 	mime_map["js"] =  	"text/ecmascript";
-// 	mime_map["jut"] =  	"image/jutvision";
-// 	mime_map["kar"] =  	"audio/midi";
-// 	mime_map["kar"] =  	"music/x-karaoke";
-// 	mime_map["ksh"] =  	"application/x-ksh";
-// 	mime_map["ksh"] =  	"text/x-script.ksh";
-// 	mime_map["la"] =  	"audio/nspaudio";
-// 	mime_map["la"] =  	"audio/x-nspaudio";
-// 	mime_map["lam"] =  	"audio/x-liveaudio";
-// 	mime_map["latex"] =  	"application/x-latex";
-// 	mime_map["lha"] =  	"application/lha";
-// 	mime_map["lha"] =  	"application/octet-stream";
-// 	mime_map["lha"] =  	"application/x-lha";
-// 	mime_map["lhx"] =  	"application/octet-stream";
-// 	mime_map["list"] =  	"text/plain";
-// 	mime_map["lma"] =  	"audio/nspaudio";
-// 	mime_map["lma"] =  	"audio/x-nspaudio";
-// 	mime_map["log"] =  	"text/plain";
-// 	mime_map["lsp"] =  	"application/x-lisp";
-// 	mime_map["lsp"] =  	"text/x-script.lisp";
-// 	mime_map["lst"] =  	"text/plain";
-// 	mime_map["lsx"] =  	"text/x-la-asf";
-// 	mime_map["ltx"] =  	"application/x-latex";
-// 	mime_map["lzh"] =  	"application/octet-stream";
-// 	mime_map["lzh"] =  	"application/x-lzh";
-// 	mime_map["lzx"] =  	"application/lzx";
-// 	mime_map["lzx"] =  	"application/octet-stream";
-// 	mime_map["lzx"] =  	"application/x-lzx";
-// 	mime_map["m"] =  	"text/plain";
-// 	mime_map["m"] =  	"text/x-m";
-// 	mime_map["m1v"] =  	"video/mpeg";
-// 	mime_map["m2a"] =  	"audio/mpeg";
-// 	mime_map["m2v"] =  	"video/mpeg";
-// 	mime_map["m3u"] =  	"audio/x-mpequrl";
-// 	mime_map["man"] =  	"application/x-troff-man";
-// 	mime_map["map"] =  	"application/x-navimap";
-// 	mime_map["mar"] =  	"text/plain";
-// 	mime_map["mbd"] =  	"application/mbedlet";
-// 	mime_map["mc$"] = "application/x-magic-cap-package-1.0";
-// 	mime_map["mcd"] =  	"application/mcad";
-// 	mime_map["mcd"] =  	"application/x-mathcad";
-// 	mime_map["mcf"] =  	"image/vasa";
-// 	mime_map["mcf"] =  	"text/mcf";
-// 	mime_map["mcp"] =  	"application/netmc";
-// 	mime_map["me"] =  	"application/x-troff-me";
-// 	mime_map["mht"] =  	"message/rfc822";
-// 	mime_map["mhtml"] =  	"message/rfc822";
-// 	mime_map["mid"] =  	"application/x-midi";
-// 	mime_map["mid"] =  	"audio/midi";
-// 	mime_map["mid"] =  	"audio/x-mid";
-// 	mime_map["mid"] =  	"audio/x-midi";
-// 	mime_map["mid"] =  	"music/crescendo";
-// 	mime_map["mid"] =  	"x-music/x-midi";
-// 	mime_map["midi"] =  	"application/x-midi";
-// 	mime_map["midi"] =  	"audio/midi";
-// 	mime_map["midi"] =  	"audio/x-mid";
-// 	mime_map["midi"] =  	"audio/x-midi";
-// 	mime_map["midi"] =  	"music/crescendo";
-// 	mime_map["midi"] =  	"x-music/x-midi";
-// 	mime_map["mif"] =  	"application/x-frame";
-// 	mime_map["mif"] =  	"application/x-mif";
-// 	mime_map["mime"] =  	"message/rfc822";
-// 	mime_map["mime"] =  	"www/mime";
-// 	mime_map["mjf"] =  	"audio/x-vnd.audioexplosion.mjuicemediafile";
-// 	mime_map["mjpg"] =  	"video/x-motion-jpeg";
-// 	mime_map["mm"] =  	"application/base64";
-// 	mime_map["mm"] =  	"application/x-meme";
-// 	mime_map["mme"] =  	"application/base64";
-// 	mime_map["mod"] =  	"audio/mod";
-// 	mime_map["mod"] =  	"audio/x-mod";
-// 	mime_map["moov"] =  	"video/quicktime";
-// 	mime_map["mov"] =  	"video/quicktime";
-// 	mime_map["movie"] =  	"video/x-sgi-movie";
-// 	mime_map["mp2"] =  	"audio/mpeg";
-// 	mime_map["mp2"] =  	"audio/x-mpeg";
-// 	mime_map["mp2"] =  	"video/mpeg";
-// 	mime_map["mp2"] =  	"video/x-mpeg";
-// 	mime_map["mp2"] =  	"video/x-mpeq2a";
-// 	mime_map["mp3"] =  	"audio/mpeg3";
-// 	mime_map["mp3"] =  	"audio/x-mpeg-3";
-// 	mime_map["mp3"] =  	"video/mpeg";
-// 	mime_map["mp3"] =  	"video/x-mpeg";
-// 	mime_map["mpa"] =  	"audio/mpeg";
-// 	mime_map["mpa"] =  	"video/mpeg";
-// 	mime_map["mpc"] =  	"application/x-project";
-// 	mime_map["mpe"] =  	"video/mpeg";
-// 	mime_map["mpeg"] =  	"video/mpeg";
-// 	mime_map["mpg"] =  	"audio/mpeg";
-// 	mime_map["mpg"] =  	"video/mpeg";
-// 	mime_map["mpga"] =  	"audio/mpeg";
-// 	mime_map["mpp"] =  	"application/vnd.ms-project";
-// 	mime_map["mpt"] =  	"application/x-project";
-// 	mime_map["mpv"] =  	"application/x-project";
-// 	mime_map["mpx"] =  	"application/x-project";
-// 	mime_map["mrc"] =  	"application/marc";
-// 	mime_map["ms"] =  	"application/x-troff-ms";
-// 	mime_map["mv"] =  	"video/x-sgi-movie";
-// 	mime_map["my"] =  	"audio/make";
-// 	mime_map["mzz"] =  	"application/x-vnd.audioexplosion.mzz";
-// 	mime_map["nap"] =  	"image/naplps";
-// 	mime_map["naplps"] =  	"image/naplps";
-// 	mime_map["nc"] =  	"application/x-netcdf";
-// 	mime_map["ncm"] =  	"application/vnd.nokia.configuration-message";
-// 	mime_map["nif"] =  	"image/x-niff";
-// 	mime_map["niff"] =  	"image/x-niff";
-// 	mime_map["nix"] =  	"application/x-mix-transfer";
-// 	mime_map["nsc"] =  	"application/x-conference";
-// 	mime_map["nvd"] =  	"application/x-navidoc";
-// 	mime_map["o"] =  	"application/octet-stream";
-// 	mime_map["oda"] =  	"application/oda";
-// 	mime_map["omc"] =  	"application/x-omc";
-// 	mime_map["omcd"] =  	"application/x-omcdatamaker";
-// 	mime_map["omcr"] =  	"application/x-omcregerator";
-// 	mime_map["p"] =  	"text/x-pascal";
-// 	mime_map["p10"] =  	"application/pkcs10";
-// 	mime_map["p10"] =  	"application/x-pkcs10";
-// 	mime_map["p12"] =  	"application/pkcs-12";
-// 	mime_map["p12"] =  	"application/x-pkcs12";
-// 	mime_map["p7a"] =  	"application/x-pkcs7-signature";
-// 	mime_map["p7c"] =  	"application/pkcs7-mime";
-// 	mime_map["p7c"] =  	"application/x-pkcs7-mime";
-// 	mime_map["p7m"] =  	"application/pkcs7-mime";
-// 	mime_map["p7m"] =  	"application/x-pkcs7-mime";
-// 	mime_map["p7r"] =  	"application/x-pkcs7-certreqresp";
-// 	mime_map["p7s"] =  	"application/pkcs7-signature";
-// 	mime_map["part"] =  	"application/pro_eng";
-// 	mime_map["pas"] =  	"text/pascal";
-// 	mime_map["pbm"] =  	"image/x-portable-bitmap";
-// 	mime_map["pcl"] =  	"application/vnd.hp-pcl";
-// 	mime_map["pcl"] =  	"application/x-pcl";
-// 	mime_map["pct"] =  	"image/x-pict";
-// 	mime_map["pcx"] =  	"image/x-pcx";
-// 	mime_map["pdb"] =  	"chemical/x-pdb";
-// 	mime_map["pdf"] =  	"application/pdf";
-// 	mime_map["pfunk"] =  	"audio/make";
-// 	mime_map["pfunk"] =  	"audio/make.my.funk";
-// 	mime_map["pgm"] =  	"image/x-portable-graymap";
-// 	mime_map["pgm"] =  	"image/x-portable-greymap";
-// 	mime_map["pic"] =  	"image/pict";
-// 	mime_map["pict"] =  	"image/pict";
-// 	mime_map["pkg"] =  	"application/x-newton-compatible-pkg";
-// 	mime_map["pko"] =  	"application/vnd.ms-pki.pko";
-// 	mime_map["pl"] =  	"text/plain";
-// 	mime_map["pl"] =  	"text/x-script.perl";
-// 	mime_map["plx"] =  	"application/x-pixclscript";
-// 	mime_map["pm"] =  	"image/x-xpixmap";
-// 	mime_map["pm"] =  	"text/x-script.perl-module";
-// 	mime_map["pm4"] =  	"application/x-pagemaker";
-// 	mime_map["pm5"] =  	"application/x-pagemaker";
-// 	mime_map["png"] =  	"image/png";
-// 	mime_map["pnm"] =  	"application/x-portable-anymap";
-// 	mime_map["pnm"] =  	"image/x-portable-anymap";
-// 	mime_map["pot"] =  	"application/mspowerpoint";
-// 	mime_map["pot"] =  	"application/vnd.ms-powerpoint";
-// 	mime_map["pov"] =  	"model/x-pov";
-// 	mime_map["ppa"] =  	"application/vnd.ms-powerpoint";
-// 	mime_map["ppm"] =  	"image/x-portable-pixmap";
-// 	mime_map["pps"] =  	"application/mspowerpoint";
-// 	mime_map["pps"] =  	"application/vnd.ms-powerpoint";
-// 	mime_map["ppt"] =  	"application/mspowerpoint";
-// 	mime_map["ppt"] =  	"application/powerpoint";
-// 	mime_map["ppt"] =  	"application/vnd.ms-powerpoint";
-// 	mime_map["ppt"] =  	"application/x-mspowerpoint";
-// 	mime_map["ppz"] =  	"application/mspowerpoint";
-// 	mime_map["pre"] =  	"application/x-freelance";
-// 	mime_map["prt"] =  	"application/pro_eng";
-// 	mime_map["ps"] =  	"application/postscript";
-// 	mime_map["psd"] =  	"application/octet-stream";
-// 	mime_map["pvu"] =  	"paleovu/x-pv";
-// 	mime_map["pwz"] =  	"application/vnd.ms-powerpoint";
-// 	mime_map["py"] =  	"text/x-script.phyton";
-// 	mime_map["pyc"] =  	"application/x-bytecode.python";
-// 	mime_map["qcp"] =  	"audio/vnd.qcelp";
-// 	mime_map["qd3"] =  	"x-world/x-3dmf";
-// 	mime_map["qd3d"] =  	"x-world/x-3dmf";
-// 	mime_map["qif"] =  	"image/x-quicktime";
-// 	mime_map["qt"] =  	"video/quicktime";
-// 	mime_map["qtc"] =  	"video/x-qtc";
-// 	mime_map["qti"] =  	"image/x-quicktime";
-// 	mime_map["qtif"] =  	"image/x-quicktime";
-// 	mime_map["ra"] =  	"audio/x-pn-realaudio";
-// 	mime_map["ra"] =  	"audio/x-pn-realaudio-plugin";
-// 	mime_map["ra"] =  	"audio/x-realaudio";
-// 	mime_map["ram"] =  	"audio/x-pn-realaudio";
-// 	mime_map["ras"] =  	"application/x-cmu-raster";
-// 	mime_map["ras"] =  	"image/cmu-raster";
-// 	mime_map["ras"] =  	"image/x-cmu-raster";
-// 	mime_map["rast"] =  	"image/cmu-raster";
-// 	mime_map["rexx"] =  	"text/x-script.rexx";
-// 	mime_map["rf"] =  	"image/vnd.rn-realflash";
-// 	mime_map["rgb"] =  	"image/x-rgb";
-// 	mime_map["rm"] =  	"application/vnd.rn-realmedia";
-// 	mime_map["rm"] =  	"audio/x-pn-realaudio";
-// 	mime_map["rmi"] =  	"audio/mid";
-// 	mime_map["rmm"] =  	"audio/x-pn-realaudio";
-// 	mime_map["rmp"] =  	"audio/x-pn-realaudio";
-// 	mime_map["rmp"] =  	"audio/x-pn-realaudio-plugin";
-// 	mime_map["rng"] =  	"application/ringing-tones";
-// 	mime_map["rng"] =  	"application/vnd.nokia.ringing-tone";
-// 	mime_map["rnx"] =  	"application/vnd.rn-realplayer";
-// 	mime_map["roff"] =  	"application/x-troff";
-// 	mime_map["rp"] =  	"image/vnd.rn-realpix";
-// 	mime_map["rpm"] =  	"audio/x-pn-realaudio-plugin";
-// 	mime_map["rt"] =  	"text/richtext";
-// 	mime_map["rt"] =  	"text/vnd.rn-realtext";
-// 	mime_map["rtf"] =  	"application/rtf";
-// 	mime_map["rtf"] =  	"application/x-rtf";
-// 	mime_map["rtf"] =  	"text/richtext";
-// 	mime_map["rtx"] =  	"application/rtf";
-// 	mime_map["rtx"] =  	"text/richtext";
-// 	mime_map["rv"] =  	"video/vnd.rn-realvideo";
-// 	mime_map["s"] =  	"text/x-asm";
-// 	mime_map["s3m"] =  	"audio/s3m";
-// 	mime_map["saveme"] =  	"application/octet-stream";
-// 	mime_map["sbk"] =  	"application/x-tbook";
-// 	mime_map["scm"] =  	"application/x-lotusscreencam";
-// 	mime_map["scm"] =  	"text/x-script.guile";
-// 	mime_map["scm"] =  	"text/x-script.scheme";
-// 	mime_map["scm"] =  	"video/x-scm";
-// 	mime_map["sdml"] =  	"text/plain";
-// 	mime_map["sdp"] =  	"application/sdp";
-// 	mime_map["sdp"] =  	"application/x-sdp";
-// 	mime_map["sdr"] =  	"application/sounder";
-// 	mime_map["sea"] =  	"application/sea";
-// 	mime_map["sea"] =  	"application/x-sea";
-// 	mime_map["set"] =  	"application/set";
-// 	mime_map["sgm"] =  	"text/sgml";
-// 	mime_map["sgm"] =  	"text/x-sgml";
-// 	mime_map["sgml"] =  	"text/sgml";
-// 	mime_map["sgml"] =  	"text/x-sgml";
-// 	mime_map["sh"] =  	"application/x-bsh";
-// 	mime_map["sh"] =  	"application/x-sh";
-// 	mime_map["sh"] =  	"application/x-shar";
-// 	mime_map["sh"] =  	"text/x-script.sh";
-// 	mime_map["shar"] =  	"application/x-bsh";
-// 	mime_map["shar"] =  	"application/x-shar";
-// 	mime_map["shtml"] =  	"text/html";
-// 	mime_map["shtml"] =  	"text/x-server-parsed-html";
-// 	mime_map["sid"] =  	"audio/x-psid";
-// 	mime_map["sit"] =  	"application/x-sit";
-// 	mime_map["sit"] =  	"application/x-stuffit";
-// 	mime_map["skd"] =  	"application/x-koan";
-// 	mime_map["skm"] =  	"application/x-koan";
-// 	mime_map["skp"] =  	"application/x-koan";
-// 	mime_map["skt"] =  	"application/x-koan";
-// 	mime_map["sl"] =  	"application/x-seelogo";
-// 	mime_map["smi"] =  	"application/smil";
-// 	mime_map["smil"] =  	"application/smil";
-// 	mime_map["snd"] =  	"audio/basic";
-// 	mime_map["snd"] =  	"audio/x-adpcm";
-// 	mime_map["sol"] =  	"application/solids";
-// 	mime_map["spc"] =  	"application/x-pkcs7-certificates";
-// 	mime_map["spc"] =  	"text/x-speech";
-// 	mime_map["spl"] =  	"application/futuresplash";
-// 	mime_map["spr"] =  	"application/x-sprite";
-// 	mime_map["sprite"] =  	"application/x-sprite";
-// 	mime_map["src"] =  	"application/x-wais-source";
-// 	mime_map["ssi"] =  	"text/x-server-parsed-html";
-// 	mime_map["ssm"] =  	"application/streamingmedia";
-// 	mime_map["sst"] =  	"application/vnd.ms-pki.certstore";
-// 	mime_map["step"] =  	"application/step";
-// 	mime_map["stl"] =  	"application/sla";
-// 	mime_map["stl"] =  	"application/vnd.ms-pki.stl";
-// 	mime_map["stl"] =  	"application/x-navistyle";
-// 	mime_map["stp"] =  	"application/step";
-// 	mime_map["sv4cpio"] =  	"application/x-sv4cpio";
-// 	mime_map["sv4crc"] =  	"application/x-sv4crc";
-// 	mime_map["svf"] =  	"image/vnd.dwg";
-// 	mime_map["svf"] =  	"image/x-dwg";
-// 	mime_map["svr"] =  	"application/x-world";
-// 	mime_map["svr"] =  	"x-world/x-svr";
-// 	mime_map["swf"] =  	"application/x-shockwave-flash";
-// 	mime_map["t"] =  	"application/x-troff";
-// 	mime_map["talk"] =  	"text/x-speech";
-// 	mime_map["tar"] =  	"application/x-tar";
-// 	mime_map["tbk"] =  	"application/toolbook";
-// 	mime_map["tbk"] =  	"application/x-tbook";
-// 	mime_map["tcl"] =  	"application/x-tcl";
-// 	mime_map["tcl"] =  	"text/x-script.tcl";
-// 	mime_map["tcsh"] =  	"text/x-script.tcsh";
-// 	mime_map["tex"] =  	"application/x-tex";
-// 	mime_map["texi"] =  	"application/x-texinfo";
-// 	mime_map["texinfo"] =  	"application/x-texinfo";
-// 	mime_map["text"] =  	"application/plain";
-// 	mime_map["text"] =  	"text/plain";
-// 	mime_map["tgz"] =  	"application/gnutar";
-// 	mime_map["tgz"] =  	"application/x-compressed";
-// 	mime_map["tif"] =  	"image/tiff";
-// 	mime_map["tif"] =  	"image/x-tiff";
-// 	mime_map["tiff"] =  	"image/tiff";
-// 	mime_map["tiff"] =  	"image/x-tiff";
-// 	mime_map["tr"] =  	"application/x-troff";
-// 	mime_map["tsi"] =  	"audio/tsp-audio";
-// 	mime_map["tsp"] =  	"application/dsptype";
-// 	mime_map["tsp"] =  	"audio/tsplayer";
-// 	mime_map["tsv"] =  	"text/tab-separated-values";
-// 	mime_map["turbot"] =  	"image/florian";
-// 	mime_map["txt"] =  	"text/plain";
-// 	mime_map["uil"] =  	"text/x-uil";
-// 	mime_map["uni"] =  	"text/uri-list";
-// 	mime_map["unis"] =  	"text/uri-list";
-// 	mime_map["unv"] =  	"application/i-deas";
-// 	mime_map["uri"] =  	"text/uri-list";
-// 	mime_map["uris"] =  	"text/uri-list";
-// 	mime_map["ustar"] =  	"application/x-ustar";
-// 	mime_map["ustar"] =  	"multipart/x-ustar";
-// 	mime_map["uu"] =  	"application/octet-stream";
-// 	mime_map["uu"] =  	"text/x-uuencode";
-// 	mime_map["uue"] =  	"text/x-uuencode";
-// 	mime_map["vcd"] =  	"application/x-cdlink";
-// 	mime_map["vcs"] =  	"text/x-vcalendar";
-// 	mime_map["vda"] =  	"application/vda";
-// 	mime_map["vdo"] =  	"video/vdo";
-// 	mime_map["vew"] =  	"application/groupwise";
-// 	mime_map["viv"] =  	"video/vivo";
-// 	mime_map["viv"] =  	"video/vnd.vivo";
-// 	mime_map["vivo"] =  	"video/vivo";
-// 	mime_map["vivo"] =  	"video/vnd.vivo";
-// 	mime_map["vmd"] =  	"application/vocaltec-media-desc";
-// 	mime_map["vmf"] =  	"application/vocaltec-media-file";
-// 	mime_map["voc"] =  	"audio/voc";
-// 	mime_map["voc"] =  	"audio/x-voc";
-// 	mime_map["vos"] =  	"video/vosaic";
-// 	mime_map["vox"] =  	"audio/voxware";
-// 	mime_map["vqe"] =  	"audio/x-twinvq-plugin";
-// 	mime_map["vqf"] =  	"audio/x-twinvq";
-// 	mime_map["vql"] =  	"audio/x-twinvq-plugin";
-// 	mime_map["vrml"] =  	"application/x-vrml";
-// 	mime_map["vrml"] =  	"model/vrml";
-// 	mime_map["vrml"] =  	"x-world/x-vrml";
-// 	mime_map["vrt"] =  	"x-world/x-vrt";
-// 	mime_map["vsd"] =  	"application/x-visio";
-// 	mime_map["vst"] =  	"application/x-visio";
-// 	mime_map["vsw"] =  	"application/x-visio";
-// 	mime_map["w60"] =  	"application/wordperfect6.0";
-// 	mime_map["w61"] =  	"application/wordperfect6.1";
-// 	mime_map["w6w"] =  	"application/msword";
-// 	mime_map["wav"] =  	"audio/wav";
-// 	mime_map["wav"] =  	"audio/x-wav";
-// 	mime_map["wb1"] =  	"application/x-qpro";
-// 	mime_map["wbmp"] =  	"image/vnd.wap.wbmp";
-// 	mime_map["web"] =  	"application/vnd.xara";
-// 	mime_map["wiz"] =  	"application/msword";
-// 	mime_map["wk1"] =  	"application/x-123";
-// 	mime_map["wmf"] =  	"windows/metafile";
-// 	mime_map["wml"] =  "text/vnd.wap.wml";
-// 	mime_map["wmlc"] = 	"application/vnd.wap.wmlc";
-// 	mime_map["wmls"] = 	"text/vnd.wap.wmlscript";
-// 	mime_map["wmlsc"] = 	"application/vnd.wap.wmlscriptc";
-// 	mime_map["word"] = 	"application/msword";
-// 	mime_map["wp"] = "application/wordperfect";
-// 	mime_map["wp5"] = "application/wordperfect";
-// 	mime_map["wp5"] = "application/wordperfect6.0";
-// 	mime_map["wp6"] = "application/wordperfect";
-// 	mime_map["wpd"] = "application/wordperfect";
-// 	mime_map["wpd"] = "application/x-wpwin";
-// 	mime_map["wq1"] = "application/x-lotus";
-// 	mime_map["wri"] = "application/mswrite";
-// 	mime_map["wri"] = "application/x-wri";
-// 	mime_map["wrl"] = "application/x-world";
-// 	mime_map["wrl"] = "model/vrml";
-// 	mime_map["wrl"] = "x-world/x-vrml";
-// 	mime_map["wrz"] = "model/vrml";
-// 	mime_map["wrz"] = "x-world/x-vrml";
-// 	mime_map["wsc"] = "text/scriplet";
-// 	mime_map["wsrc"] = "application/x-wais-source";
-// 	mime_map["wtk"] = "application/x-wintalk";
-// 	mime_map["xbm"] = "image/x-xbitmap";
-// 	mime_map["xbm"] = "image/x-xbm";
-// 	mime_map["xbm"] = "image/xbm";
-// 	mime_map["xdr"] = "video/x-amt-demorun";
-// 	mime_map["xgz"] = "xgl/drawing";
-// 	mime_map["xif"] = "image/vnd.xiff";
-// 	mime_map["xl"] = "application/excel";
-// 	mime_map["xla"] = "application/excel";
-// 	mime_map["xla"] = "application/x-excel";
-// 	mime_map["xla"] = "application/x-msexcel";
-// 	mime_map["xlb"] = "application/excel";
-// 	mime_map["xlb"] = "application/vnd.ms-excel";
-// 	mime_map["xlb"] = "application/x-excel";
-// 	mime_map["xlc"] = "application/excel";
-// 	mime_map["xlc"] = "application/vnd.ms-excel";
-// 	mime_map["xlc"] = "application/x-excel";
-// 	mime_map["xld"] = "application/excel";
-// 	mime_map["xld"] = "application/x-excel";
-// 	mime_map["xlk"] = "application/excel";
-// 	mime_map["xlk"] = "application/x-excel";
-// 	mime_map["xll"] = "application/excel";
-// 	mime_map["xll"] = "application/vnd.ms-excel";
-// 	mime_map["xll"] = "application/x-excel";
-// 	mime_map["xlm"] = "application/excel";
-// 	mime_map["xlm"] = "application/vnd.ms-excel";
-// 	mime_map["xlm"] = "application/x-excel";
-// 	mime_map["xls"] = "application/excel";
-// 	mime_map["xls"] = "application/vnd.ms-excel";
-// 	mime_map["xls"] = "application/x-excel";
-// 	mime_map["xls"] = "application/x-msexcel";
-// 	mime_map["xlt"] = "application/excel";
-// 	mime_map["xlt"] = "application/x-excel";
-// 	mime_map["xlv"] = "application/excel";
-// 	mime_map["xlv"] = "application/x-excel";
-// 	mime_map["xlw"] = "application/excel";
-// 	mime_map["xlw"] = "application/vnd.ms-excel";
-// 	mime_map["xlw"] = "application/x-excel";
-// 	mime_map["xlw"] = "application/x-msexcel";
-// 	mime_map["xm"] = "audio/xm";
-// 	mime_map["xml"] = "application/xml";
-// 	mime_map["xml"] = "text/xml";
-// 	mime_map["xmz"] = "xgl/movie";
-// 	mime_map["xpix"] = "application/x-vnd.ls-xpix";
-// 	mime_map["xpm"] = "image/x-xpixmap";
-// 	mime_map["xpm"] = "image/xpm";
-// 	mime_map["x-png"] = "image/png";
-// 	mime_map["xsr"] = "video/x-amt-showrun";
-// 	mime_map["xwd"] = "image/x-xwd";
-// 	mime_map["xwd"] = "image/x-xwindowdump";
-// 	mime_map["xyz"] = "chemical/x-pdb";
-// 	mime_map["z"] = "application/x-compress";
-// 	mime_map["z"] = "application/x-compressed";
-// 	mime_map["zip"] = "application/x-compressed";
-// 	mime_map["zip"] = "application/x-zip-compressed";
-// 	mime_map["zip"] = "application/zip";
-// 	mime_map["zip"] = "multipart/x-zip";
-// 	mime_map["zoo"] = "application/octet-stream";
-// 	mime_map["zsh"] = "text/x-script.zsh";
-// 	return mime_map;
-// }
-
-} // namespace webserv
