+++ tests/test_HTTPRequest.cpp
@@ -1,7 +1,7 @@
-#include "HTTPRequest.hpp"
+#include "http/HTTPRequest.hpp"
 
 #include <gtest/gtest.h>
 
 TEST(test_HTTPRequest, constructor) {
-    webserv::HTTPRequest req(std::string("hello"));
+    HTTPRequest req(std::string("hello"));
 }
